[crypto] Split ECC API into four headers.
Code is only moved in this commit, not changed.

This will later allow us to easily exclude unnecessary elliptic-curve suites
from the cryptolib binary.
